src/solvers/pattern_solver.py
# src/solvers/pattern_solver.py

import re
import logging
from typing import Dict, List, Optional, Tuple, Any
import math
from functools import reduce
from dataclasses import dataclass

from sympy import sympify, SympifyError
from sympy.parsing.latex import parse_latex

from .base_solver import BaseSolver, MathProblem

logger = logging.getLogger(__name__)

@dataclass
class PatternMatch:
    """Structure to hold pattern matching results"""
    pattern_type: str
    matches: List[str]
    confidence: float  # 0.0 to 1.0
    extracted_values: List[Any] = None

class PatternBasedSolver(BaseSolver):
    """
    A solver that uses pattern recognition to identify and solve mathematical problems.
    Implements various pattern matching strategies and specialized solution methods.
    """
    
    def __init__(self):
        super().__init__()
        self._initialize_patterns()
    
    def _initialize_patterns(self):
        """Initialize pattern matching rules and their handlers"""
        # Mathematical patterns
        self.math_patterns = {
            'modulo': {
                'pattern': r'modulo\s+(\d+)|mod\s+(\d+)',
                'handler': self._handle_modulo,
                'priority': 1
            },
            'factorial': {
                'pattern': r'(\d+)!',
                'handler': self._handle_factorial,
                'priority': 2
            },
            'sequence': {
                'pattern': r'(a_n|F_n|n\s*th\s*term)',
                'handler': self._handle_sequence,
                'priority': 2
            },
            'divisibility': {
                'pattern': r'divisible|multiple|factor',
                'handler': self._handle_divisibility,
                'priority': 1
            },
            'geometry': {
                'pattern': r'triangle|circle|angle|perpendicular',
                'handler': self._handle_geometry,
                'priority': 3
            }
        }
        
        # Problem type indicators
        self.type_indicators = {
            'number_theory': [
                'prime', 'factor', 'divisible', 'modulo', 'remainder',
                'coprime', 'gcd', 'lcm'
            ],
            'geometry': [
                'triangle', 'circle', 'angle', 'perpendicular', 'parallel',
                'distance', 'area', 'volume'
            ],
            'algebra': [
                'equation', 'solve', 'polynomial', 'expression', 'simplify',
                'evaluate'
            ],
            'combinatorics': [
                'ways', 'arrange', 'permutation', 'combination', 'choose',
                'possible'
            ],
            'sequence': [
                'sequence', 'series', 'fibonacci', 'arithmetic', 'geometric',
                'term', 'recursive'
            ]
        }
    
    def solve(self, question: str) -> int:
        """
        Main solving method that implements pattern-based problem solving.
        
        Args:
            question: The LaTeX formatted problem text
            
        Returns:
            int: Solution modulo 1000
        """
        try:
            # Parse problem
            problem = self.preprocess(question)
            
            # Find all matching patterns
            matches = self._find_patterns(problem)
            
            # Sort matches by priority and confidence
            sorted_matches = sorted(
                matches,
                key=lambda x: (
                    self.math_patterns[x.pattern_type]['priority'],
                    x.confidence
                ),
                reverse=True
            )
            
            # Try patterns in order
            for match in sorted_matches:
                handler = self.math_patterns[match.pattern_type]['handler']
                try:
                    result = handler(problem, match)
                    if result is not None:
                        return self._ensure_valid_answer(result)
                except Exception as e:
                    logger.debug(f"Handler {match.pattern_type} failed: {e}")
                    continue
            
            # Fallback to type-specific general handlers
            result = self._handle_by_type(problem)
            if result is not None:
                return self._ensure_valid_answer(result)
            
            # Last resort: basic numerical analysis
            return self._basic_numerical_analysis(problem)
            
        except Exception as e:
            logger.error(f"Error in pattern solver: {e}")
            return 42
    
    def _find_patterns(self, problem: MathProblem) -> List[PatternMatch]:
        """Find all matching patterns in the problem"""
        matches = []
        for pattern_type, pattern_info in self.math_patterns.items():
            pattern = pattern_info['pattern']
            found_matches = re.finditer(pattern, problem.cleaned_text, re.IGNORECASE)
            
            for match in found_matches:
                confidence = self._calculate_pattern_confidence(
                    pattern_type, match.group(), problem
                )
                matches.append(PatternMatch(
                    pattern_type=pattern_type,
                    matches=[match.group()],
                    confidence=confidence,
                    extracted_values=self._extract_pattern_values(match)
                ))
        
        return matches
    
    def _calculate_pattern_confidence(
        self, pattern_type: str, match: str, problem: MathProblem
    ) -> float:
        """Calculate confidence score for a pattern match"""
        # Base confidence from match presence
        confidence = 0.5
        
        # Adjust based on context
        if any(indicator in problem.cleaned_text.lower() 
               for indicator in self.type_indicators.get(pattern_type, [])):
            confidence += 0.2
            
        # Adjust based on mathematical expressions
        if any(match in expr for expr in problem.expressions):
            confidence += 0.2
            
        # Adjust based on problem type
        if problem.problem_type == pattern_type:
            confidence += 0.1
            
        return min(confidence, 1.0)
    
    def _extract_pattern_values(self, match) -> List[Any]:
        """Extract numerical or other values from pattern match"""
        values = []
        for group in match.groups():
            if group is not None:
                try:
                    # Try to convert to number
                    values.append(float(group))
                except ValueError:
                    # Keep as string if not a number
                    values.append(group)
        return values
    
    def _handle_modulo(self, problem: MathProblem, match: PatternMatch) -> Optional[int]:
        """Handle modulo operation patterns"""
        try:
            modulo = match.extracted_values[0]
            numbers = problem.numbers
            
            if not numbers:
                return None
                
            # Try to find the most relevant number for modulo
            result = numbers[-1] if len(numbers) > 0 else 0
            return int(result) % int(modulo)
        except Exception:
            return None
    
    def _handle_factorial(self, problem: MathProblem, match: PatternMatch) -> Optional[int]:
        """Handle factorial patterns"""
        try:
            n = int(match.extracted_values[0])
            if n > 20:  # Avoid large factorials
                return self._factorial_mod_1000(n)
            return math.factorial(n) % 1000
        except Exception:
            return None
    
    def _handle_sequence(self, problem: MathProblem, match: PatternMatch) -> Optional[int]:
        """Handle sequence patterns"""
        try:
            if 'fibonacci' in problem.cleaned_text.lower():
                return self._handle_fibonacci(problem)
            if 'arithmetic' in problem.cleaned_text.lower():
                return self._handle_arithmetic_sequence(problem)
            return None
        except Exception:
            return None
    
    def _handle_divisibility(self, problem: MathProblem, match: PatternMatch) -> Optional[int]:
        """Handle divisibility patterns"""
        try:
            numbers = problem.numbers
            if len(numbers) >= 2:
                return self._gcd(int(numbers[0]), int(numbers[1])) % 1000
            return None
        except Exception:
            return None
    
    def _handle_geometry(self, problem: MathProblem, match: PatternMatch) -> Optional[int]:
        """Handle geometry patterns"""
        try:
            if 'triangle' in problem.cleaned_text.lower():
                return self._handle_triangle(problem)
            if 'circle' in problem.cleaned_text.lower():
                return self._handle_circle(problem)
            return None
        except Exception:
            return None
    
    def _handle_by_type(self, problem: MathProblem) -> Optional[int]:
        """Handle problems based on their identified type"""
        type_handlers = {
            'number_theory': self._handle_number_theory,
            'geometry': self._handle_geometry,
            'algebra': self._handle_algebra,
            'combinatorics': self._handle_combinatorics,
            'sequence': self._handle_sequence
        }
        
        handler = type_handlers.get(problem.problem_type)
        if handler:
            try:
                return handler(problem, None)
            except Exception as e:
                logger.debug(f"Type handler failed: {e}")
        return None
    
    def _basic_numerical_analysis(self, problem: MathProblem) -> int:
        """Basic numerical analysis as last resort"""
        try:
            numbers = problem.numbers
            if not numbers:
                return 42
            
            # Try common operations
            operations = [
                lambda x: max(x),
                lambda x: min(x),
                lambda x: sum(x),
                lambda x: reduce(lambda a, b: a * b, x),
            ]
            
            for op in operations:
                try:
                    result = op(numbers)
                    if 0 <= result <= 999:
                        return int(result)
                except Exception:
                    continue
            
            return int(max(numbers)) % 1000
            
        except Exception:
            return 42
    
    @staticmethod
    def _factorial_mod_1000(n: int) -> int:
        """Calculate factorial modulo 1000 for large numbers"""
        result = 1
        for i in range(1, min(n + 1, 1000)):
            result = (result * i) % 1000
        return result
    
    @staticmethod
    def _gcd(a: int, b: int) -> int:
        """Calculate greatest common divisor"""
        while b:
            a, b = b, a % b
        return a
    
    def _handle_arithmetic_sequence(self, problem: MathProblem) -> Optional[int]:
        """Handle arithmetic sequence problems"""
        try:
            numbers = problem.numbers
            if len(numbers) >= 2:
                diff = numbers[1] - numbers[0]
                return int(diff) % 1000
            return None
        except Exception:
            return None
    
    def _handle_fibonacci(self, problem: MathProblem) -> Optional[int]:
        """Handle Fibonacci sequence problems"""
        def fib_mod_1000(n: int) -> int:
            if n <= 1:
                return n
            a, b = 0, 1
            for _ in range(2, n + 1):
                a, b = b, (a + b) % 1000
            return b
        
        try:
            numbers = [n for n in problem.numbers if n <= 100]
            if numbers:
                return fib_mod_1000(int(min(numbers)))
            return None
        except Exception:
            return None

    def __str__(self) -> str:
        """String representation"""
        return f"PatternBasedSolver(patterns={len(self.math_patterns)})"